Refactor Auth/Sign In
Use Axios instead of Fetch.
Create useSessions and usePostSession custom hooks.
Rename SignIn component to SignInForm.
