+++ app/javascript/components/sessions/AuthContext.jsx
@@ -1,28 +1,15 @@
 import React, { useContext, useMemo } from 'react';
-import { useQuery } from 'react-query';
+import useSessions from '../../hooks/queries/sessions/useSessions';
 
 const AuthContext = React.createContext();
 
-// A component that imports the useAuth method will be wrapped with AuthContext and thus
-// should have access to the object(s) in AuthProvider. Is this correct? - SC
 export function useAuth() {
   return useContext(AuthContext);
 }
 
 export default function AuthProvider({ children }) {
-  async function fetchCurrentUser() {
-    const response = await fetch('/api/v1/sessions.json', {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json',
-      },
-    });
-    if (!response.ok) throw new Error('User is not signed in.');
-    return response.json();
-  }
 
-  const { data: session, status, error } = useQuery('current_user', fetchCurrentUser);
+  const { data: session, status, error } = useSessions();
 
   const currentUser = {
     name: session?.current_user?.name,
@@ -31,7 +18,6 @@ export default function AuthProvider({ children }) {
     signed_in: session?.signed_in ?? false,
   };
 
-  // ESlint is suggesting to useMemo so currentUser is not changed every render. To be reviewed - SC
   const memoizedCurrentUser = useMemo(() => currentUser, [currentUser]);
 
   if (status === 'loading') return <p> Loading... </p>
