+++ app/javascript/hooks/queries/sessions/useSessions.jsx
@@ -1,0 +1,16 @@
+// TODO - samuel: Can't rename this file to useSession for some reason.
+
+import React from 'react';
+import { useQuery } from 'react-query';
+import axios from 'axios';
+
+export default function useSessions() {
+  return useQuery('currentUser', () => axios.get('/api/v1/sessions.json', {
+    headers: {
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+    },
+  })
+    .then((resp) => resp.data)
+    .catch((error) => console.log(error)));
+}
