+++ app/javascript/main.jsx
@@ -6,7 +6,7 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import {QueryClient, QueryClientProvider} from "react-query";
 import App from "./app";
 import Signup from "./routes/Signup";
-import SignIn from './components/sessions/SignIn';
+import SignInForm from './components/sessions/SignInForm';
 import AuthProvider from './components/sessions/AuthContext';
 import RoomView from "./components/rooms/RoomView";
 import Rooms from "./routes/Rooms"
@@ -21,7 +21,7 @@ const root = (
                  <Route path="/" element={<App />}>
                    <Route index element={<h1 className="text-center">Index</h1>} />
                    <Route path="/signup" element={<Signup />} />
-                   <Route path="/sign_in" element={<SignIn />} />
+                   <Route path="/sign_in" element={<SignInForm />} />
                    <Route path="/rooms" element={<Rooms/>} />
                    <Route path="/rooms/:friendly_id" element={<RoomView />} />
                    <Route path="*" element={<h1 className="text-center">404</h1>} />
